+++ notification-service/src/main/java/com/ecommerce/notificationservice/service/PaymentService.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface PaymentService {
     Mono<Payment> savePayment(PaymentDto paymentDto);
-    Mono<Payment> getPayment(Integer paymentId);
+    Mono<Payment> getPayment(Long paymentId);
     Mono<List<Payment>> getAllPayments();
-    Mono<Void> deletePayment(Integer paymentId);
+    Mono<Void> deletePayment(Long paymentId);
 }
